+++ branches/kirk/SeaBreeze/src/api/seabreezeapi/OpticalBenchFeatureAdapter.cpp
@@ -60,8 +60,8 @@ OpticalBenchFeatureAdapter::~OpticalBenchFeatureAdapter() {
 #endif
 
 
-unsigned int OpticalBenchFeatureAdapter::readOpticalBenchFiberDiameterMicrons(int *errorCode) {
-	unsigned int returnValue;
+unsigned short int OpticalBenchFeatureAdapter::readOpticalBenchFiberDiameterMicrons(int *errorCode) {
+	unsigned short int returnValue;
 	
 	// no memory allocated, just pass it through
 	try {
@@ -75,8 +75,8 @@ unsigned int OpticalBenchFeatureAdapter::readOpticalBenchFiberDiameterMicrons(in
     return returnValue;
 }
 
-unsigned int OpticalBenchFeatureAdapter::readOpticalBenchSlitWidthMicrons(int *errorCode) {
-	unsigned int returnValue;
+unsigned short int OpticalBenchFeatureAdapter::readOpticalBenchSlitWidthMicrons(int *errorCode) {
+	unsigned short int returnValue;
 	
 	// no memory allocated, just pass it through
 	try {
