+++ branches/kirk/SeaBreeze/src/vendors/OceanOptics/features/optical_bench/OpticalBenchFeature.cpp
@@ -57,10 +57,10 @@ OpticalBenchFeature::~OpticalBenchFeature() {
 #endif
 
 
-unsigned int OpticalBenchFeature::readOpticalBenchFiberDiameterMicrons(const Protocol &protocol, const Bus &bus) throw (FeatureException) {
+unsigned short int OpticalBenchFeature::readOpticalBenchFiberDiameterMicrons(const Protocol &protocol, const Bus &bus) throw (FeatureException) {
 
     OpticalBenchProtocolInterface *opticalBenchPI = NULL;
-	unsigned int fiberDiameter;
+	unsigned short int fiberDiameter;
     ProtocolHelper *proto = NULL;
 
     try {
@@ -84,10 +84,10 @@ unsigned int OpticalBenchFeature::readOpticalBenchFiberDiameterMicrons(const Pro
     }
 }
 
-unsigned int OpticalBenchFeature::readOpticalBenchSlitWidthMicrons(const Protocol &protocol, const Bus &bus) throw (FeatureException) {
+unsigned short int OpticalBenchFeature::readOpticalBenchSlitWidthMicrons(const Protocol &protocol, const Bus &bus) throw (FeatureException) {
 
     OpticalBenchProtocolInterface *opticalBenchPI = NULL;
-	unsigned int slitWidth;
+	unsigned short int slitWidth;
     ProtocolHelper *proto = NULL;
 
     try {
